feat(docstoc): multi level toc
